fix: use the correct ClassLoader
When looking for HttpClient classes (e.g.
org.apache.http.client.utils.URIBuilder), use the current thread's class
loader, rather than the default.
