+++ agent/test/http_client/appmap.yml
@@ -1,3 +1,3 @@
 name: HTTP client test
 packages:
-  - path: http_client
+  - path: httpclient
