+++ agent/test/http_client/httpclient/settings.gradle
