+++ agent/test/http_client/springboot/build.gradle
@@ -1,0 +1,35 @@
+
+plugins {
+  id 'java'
+  id 'org.springframework.boot' version "${SPRING_BOOT_VERSION}"
+  id 'io.spring.dependency-management' version '1.1.4'
+
+}
+
+repositories {
+  mavenCentral()
+}
+
+sourceSets {
+  main {
+    java {
+      srcDirs = ['src/main/java']
+    }
+  }
+}
+
+dependencies {
+  implementation 'org.apache.httpcomponents:httpclient:4.5.14'
+  implementation 'org.springframework.boot:spring-boot-starter'
+}
+
+bootJar {
+  // I don't understand why hasProperty doesn't work here, but properties.containsKey does....
+  // if (hasProperty("mainClass")) {
+  if (properties.containsKey("mainClass")) {
+    mainClass = 'com.example.App'
+  }
+  else {
+    mainClassName = 'com.example.App'
+  }
+}
