+++ agent/test/http_client/springboot/src/main/java/com/example/App.java
@@ -1,0 +1,23 @@
+package com.example;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.CommandLineRunner;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class App implements CommandLineRunner {
+  @Autowired
+  private HttpClientService httpClientService;
+
+  public static void main(String[] args) {
+    SpringApplication.run(App.class, args);
+  }
+
+  @Override
+  public void run(String... args) throws Exception {
+    String url = args[0];
+    System.exit(httpClientService.get(url) ? 0 : 1);
+  }
+
+}
