+++ agent/test/http_client/springboot/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+logging.level.root=warn
+spring.main.banner-mode=off
